requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
djangorestframework==3.16.1
filelock==3.19.1
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ollama==0.6.0
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.2.0
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==78.1.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
